+++ examples/csharp-winforms/OcrResult.cs
@@ -10,6 +10,10 @@ namespace LwOpenCVDnnPPOCRWin7Test
 
     public sealed class OcrItem
     {
+        public int index { get; set; }
+        public int id { get; set; }
+        public int code { get; set; }
+        public string msg { get; set; }
         public string text { get; set; }
         public float score { get; set; }
         public int x1 { get; set; }
@@ -20,6 +24,6 @@ namespace LwOpenCVDnnPPOCRWin7Test
         public int y3 { get; set; }
         public int x4 { get; set; }
         public int y4 { get; set; }
+        public int rotation { get; set; }
     }
 }
-